Restructured files. Started work on playr class (that will actually play the music files).
In an attempt to keep the playr app clean, all class files have been moved to a "lib" folder.
A new runner file has been created as well. This fill includes all the files the apps needs. It also holds (for now) the Playr class which will actually play/pause/stop the queue.

I have also included some sample code to fork off a process (so the Sinatra app will actually run).

This playr class should probably be moved to another file within the lib file, but for now it's fine in the runner.