Fixed browse issues when album name was changed.
An issue was popping up when an album name changed. Because album names were being stored in redis, when an album name changed, it still thought those albums were around, and the new album wouldn't show up. To fix this I removed redis for storing information, and am using it only as a caching solution.

Other changes:
* added an "Add to Queue" button on the track info page.
* moved browse to a submenu (artist, album, and all)