Playr actually plays tracks now.
Did more work with the Playr class and process. It now plays the next queued track or a random track. Still need to do a little bit more validating to make sure the track hasn't been played recently.

Added a history table to store recent plays. This is also being linked to the "Now Playing" section of the queue page.

Added more api endpoints to work with the playing track (next, volume).