Process management enhancements.
Fixed some issues with process management. Terminate singals are now being handled by the WebSocket and Play processes.

The Queue database class was clashing with the thread Queue class. The Queue database class is now SongQueue. Changes have been made in both app.rb and playr.rb to reflect this change.

Added a check for empty songs before trying to play an empty song. This was causing the Playr process to crash.